Fix plotting for ICRA Tutorial 3
